reduced size of Account class by factoring out enums and helpers
