+++ myExpenses/src/androidTest/java/org/totschnig/myexpenses/test/model/ExportTest.java
@@ -65,7 +65,7 @@ public class ExportTest extends ModelTest  {
   private void insertData1() {
     Transaction op;
     account1 = new Account("Account 1",openingBalance,"Account 1");
-    account1.type = Account.Type.BANK;
+    account1.type = AccountType.BANK;
     account1.save();
     account2 = new Account("Account 2",openingBalance,"Account 2");
     account2.save();
@@ -193,7 +193,7 @@ public class ExportTest extends ModelTest  {
     };
     try {
       insertData1();
-      Result result = exportAll(account1, Account.ExportFormat.QIF, false);
+      Result result = exportAll(account1, ExportFormat.QIF, false);
       assertTrue(result.success);
       export = (Uri) result.extra[0];
       compare(new File(export.getPath()),linesQIF);
@@ -222,7 +222,7 @@ public class ExportTest extends ModelTest  {
     };
     try {
       insertData1();
-      Result result = exportAll(account1, Account.ExportFormat.CSV, false);
+      Result result = exportAll(account1, ExportFormat.CSV, false);
       assertTrue(result.success);
       export = (Uri) result.extra[0];
       compare(new File(export.getPath()),linesCSV);
@@ -250,7 +250,7 @@ public class ExportTest extends ModelTest  {
     };
     try {
       insertData1();
-      Result result = account1.exportWithFilter(outDir, FILE_NAME, Account.ExportFormat.CSV, false,
+      Result result = account1.exportWithFilter(outDir, FILE_NAME, ExportFormat.CSV, false,
           "M/d/yyyy", ',', "UTF-8", null);
       assertTrue(result.success);
       export = (Uri) result.extra[0];
@@ -268,14 +268,14 @@ public class ExportTest extends ModelTest  {
     };
     try {
       insertData1();
-      Result result = exportAll(account1,Account.ExportFormat.CSV, false);
+      Result result = exportAll(account1, ExportFormat.CSV, false);
       assertTrue("Export failed with message: " + getContext().getString(result.getMessage()),result.success);
       account1.markAsExported(null);
       export = (Uri) result.extra[0];
       //noinspection ResultOfMethodCallIgnored
       new File(export.getPath()).delete();
       insertData2();
-      result = exportAll(account1,Account.ExportFormat.CSV, true);
+      result = exportAll(account1, ExportFormat.CSV, true);
       assertTrue("Export failed with message: " + getContext().getString(result.getMessage()),result.success);
       export = (Uri) result.extra[0];
       compare(new File(export.getPath()),linesCSV);
@@ -329,7 +329,7 @@ public class ExportTest extends ModelTest  {
     }
   }
   /**
-   * calls {@link Account#exportWithFilter(DocumentFile, String, Account.ExportFormat, boolean, String, char, String, WhereFilter)} with
+   * calls {@link Account#exportWithFilter(DocumentFile, String, ExportFormat, boolean, String, char, String, WhereFilter)} with
    * * fileName TEST
    * * date format "dd/MM/yyyy"
    * * encoding UTF-8
@@ -337,7 +337,7 @@ public class ExportTest extends ModelTest  {
    * * WhereFilter null
    * should only be used from unit tests
    */
-  private Result exportAll(Account account, Account.ExportFormat format, boolean notYetExportedP) throws IOException {
+  private Result exportAll(Account account, ExportFormat format, boolean notYetExportedP) throws IOException {
     return account.exportWithFilter(outDir, FILE_NAME, format, notYetExportedP, "dd/MM/yyyy", '.', "UTF-8", null);
   }
 }
