+++ myExpenses/src/main/java/org/totschnig/myexpenses/fragment/CsvImportDataFragment.java
@@ -36,7 +36,7 @@ import org.totschnig.myexpenses.activity.ProtectionDelegate;
 import org.totschnig.myexpenses.dialog.ConfirmationDialogFragment;
 import org.totschnig.myexpenses.dialog.ProgressDialogFragment;
 import org.totschnig.myexpenses.export.qif.QifDateFormat;
-import org.totschnig.myexpenses.model.Account;
+import org.totschnig.myexpenses.model.AccountType;
 import org.totschnig.myexpenses.preference.PrefKey;
 import org.totschnig.myexpenses.task.TaskExecutionFragment;
 import org.totschnig.myexpenses.util.AcraHelper;
@@ -412,7 +412,7 @@ public class CsvImportDataFragment extends Fragment {
           long accountId = ((CsvImportActivity) getActivity()).getAccountId();
           String currency = ((CsvImportActivity) getActivity()).getCurrency();
           QifDateFormat format = ((CsvImportActivity) getActivity()).getDateFormat();
-          Account.Type type = ((CsvImportActivity) getActivity()).getAccountType();
+          AccountType type = ((CsvImportActivity) getActivity()).getAccountType();
           TaskExecutionFragment taskExecutionFragment =
               TaskExecutionFragment.newInstanceCSVImport(
                   mDataset,columnToFieldMap,discardedRows,format,accountId,currency,type);
