+++ myExpenses/src/main/java/org/totschnig/myexpenses/model/Account.java
@@ -18,14 +18,10 @@ package org.totschnig.myexpenses.model;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.math.BigDecimal;
-import java.text.Collator;
 import java.text.DateFormat;
 import java.text.DecimalFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Calendar;
-import java.util.Comparator;
 import java.util.Currency;
 import java.util.Date;
 import java.util.HashMap;
@@ -35,7 +31,6 @@ import org.totschnig.myexpenses.MyApplication;
 import org.totschnig.myexpenses.R;
 import org.totschnig.myexpenses.fragment.TransactionList;
 import org.totschnig.myexpenses.model.Transaction.CrStatus;
-import org.totschnig.myexpenses.preference.PrefKey;
 import org.totschnig.myexpenses.provider.DatabaseConstants;
 import org.totschnig.myexpenses.provider.DbUtils;
 import org.totschnig.myexpenses.provider.TransactionProvider;
@@ -54,21 +49,16 @@ import com.itextpdf.text.pdf.PdfPTable;
 import com.itextpdf.text.pdf.PdfWriter;
 import com.itextpdf.text.pdf.draw.LineSeparator;
 
-import android.annotation.TargetApi;
 import android.content.ContentProviderOperation;
 import android.content.ContentUris;
 import android.content.ContentValues;
-import android.content.Context;
 import android.content.OperationApplicationException;
 import android.database.Cursor;
 import android.net.Uri;
 import android.net.Uri.Builder;
-import android.os.Build;
 import android.os.RemoteException;
 import android.support.annotation.VisibleForTesting;
 import android.support.v4.provider.DocumentFile;
-import android.telephony.TelephonyManager;
-import android.text.TextUtils;
 import android.util.Log;
 
 import static org.totschnig.myexpenses.provider.DatabaseConstants.*;
@@ -142,444 +132,16 @@ public class Account extends Model {
     PROJECTION_FULL[baseLength + 8] = "0 AS " + KEY_IS_AGGREGATE;//this is needed in the union with the aggregates to sort real accounts first
     PROJECTION_FULL[baseLength + 9] = HAS_FUTURE;
     PROJECTION_FULL[baseLength + 10] = HAS_CLEARED;
-    PROJECTION_FULL[baseLength + 11] = Type.sqlOrderExpression();
+    PROJECTION_FULL[baseLength + 11] = AccountType.sqlOrderExpression();
     PROJECTION_FULL[baseLength + 12] = KEY_LAST_USED;
 
   }
 
   public static final Uri CONTENT_URI = TransactionProvider.ACCOUNTS_URI;
 
-  public enum ExportFormat {
-    QIF, CSV;
-
-    public String getMimeType() {
-      return "text/" + getExtension();
-    }
-
-    public String getExtension() {
-      return name().toLowerCase(Locale.US);
-    }
-  }
-
-  public enum Type {
-    CASH, BANK, CCARD, ASSET, LIABILITY;
-    public static final String JOIN;
-
-    public String toString() {
-      Context ctx = MyApplication.getInstance();
-      switch (this) {
-        case CASH:
-          return ctx.getString(R.string.account_type_cash);
-        case BANK:
-          return ctx.getString(R.string.account_type_bank);
-        case CCARD:
-          return ctx.getString(R.string.account_type_ccard);
-        case ASSET:
-          return ctx.getString(R.string.account_type_asset);
-        case LIABILITY:
-          return ctx.getString(R.string.account_type_liability);
-      }
-      return "";
-    }
-
-    public int toStringResPlural() {
-      switch (this) {
-        case CASH:
-          return R.string.account_type_cash_plural;
-        case BANK:
-          return R.string.account_type_bank_plural;
-        case CCARD:
-          return R.string.account_type_ccard_plural;
-        case ASSET:
-          return R.string.account_type_asset_plural;
-        case LIABILITY:
-          return R.string.account_type_liability_plural;
-        default:
-          return 0;
-      }
-    }
-
-    public String toQifName() {
-      switch (this) {
-        case CASH:
-          return "Cash";
-        case BANK:
-          return "Bank";
-        case CCARD:
-          return "CCard";
-        case ASSET:
-          return "Oth A";
-        case LIABILITY:
-          return "Oth L";
-      }
-      return "";
-    }
-
-    public static Type fromQifName(String qifName) {
-      if (qifName.equals("Oth L")) {
-        return LIABILITY;
-      } else if (qifName.equals("Oth A")) {
-        return ASSET;
-      } else if (qifName.equals("CCard")) {
-        return CCARD;
-      } else if (qifName.equals("Cash")) {
-        return CASH;
-      } else {
-        return BANK;
-      }
-    }
-
-    public static String sqlOrderExpression() {
-      String result = "CASE " + KEY_TYPE;
-      for (Type type : Type.values()) {
-        result += " WHEN '" + type.name() + "' THEN " + type.getSortOrder();
-      }
-      result += " ELSE -1 END AS " + KEY_SORT_KEY_TYPE;
-      return result;
-    }
-
-    private String getSortOrder() {
-      switch (this) {
-        case CASH:
-          return "0";
-        case BANK:
-          return "1";
-        case CCARD:
-          return "2";
-        case ASSET:
-          return "3";
-        case LIABILITY:
-          return "4";
-      }
-      return "-1";
-    }
-
-    static {
-      JOIN = Utils.joinEnum(Type.class);
-    }
-  }
-
-  public Type type;
-
-  /**
-   * grouping of accounts in account list
-   */
-  public enum AccountGrouping {
-    NONE, TYPE, CURRENCY
-  }
-
-  /**
-   * grouping of transactions
-   */
-  public enum Grouping {
-    NONE, DAY, WEEK, MONTH, YEAR;
-
-    /**
-     * @param groupYear   the year of the group to display
-     * @param groupSecond the number of the group in the second dimension (day, week or month)
-     * @param c           a cursor where we can find information about the current date
-     * @return a human readable String representing the group as header or activity title
-     */
-    public String getDisplayTitle(Context ctx, int groupYear, int groupSecond, Cursor c) {
-      int this_week = c.getInt(c.getColumnIndex(KEY_THIS_WEEK));
-      int this_day = c.getInt(c.getColumnIndex(KEY_THIS_DAY));
-      int this_year = c.getInt(c.getColumnIndex(KEY_THIS_YEAR));
-      Calendar cal;
-      switch (this) {
-        case NONE:
-          return ctx.getString(R.string.menu_aggregates);
-        case DAY:
-          cal = Calendar.getInstance();
-          cal.set(Calendar.YEAR, groupYear);
-          cal.set(Calendar.DAY_OF_YEAR, groupSecond);
-          String title = DateFormat.getDateInstance(DateFormat.FULL).format(cal.getTime());
-          if (groupYear == this_year) {
-            if (groupSecond == this_day)
-              return ctx.getString(R.string.grouping_today) + " (" + title + ")";
-            else if (groupSecond == this_day - 1)
-              return ctx.getString(R.string.grouping_yesterday) + " (" + title + ")";
-          }
-          return title;
-        case WEEK:
-          int this_year_of_week_start = c.getInt(c.getColumnIndex(KEY_THIS_YEAR_OF_WEEK_START));
-          DateFormat dateformat = Utils.localizedYearlessDateFormat();
-          String weekRange = " (" + Utils.convDateTime(c.getString(c.getColumnIndex(KEY_WEEK_START)), dateformat)
-              + " - " + Utils.convDateTime(c.getString(c.getColumnIndex(KEY_WEEK_END)), dateformat) + " )";
-          String yearPrefix;
-          if (groupYear == this_year_of_week_start) {
-            if (groupSecond == this_week)
-              return ctx.getString(R.string.grouping_this_week) + weekRange;
-            else if (groupSecond == this_week - 1)
-              return ctx.getString(R.string.grouping_last_week) + weekRange;
-            yearPrefix = "";
-          } else
-            yearPrefix = groupYear + ", ";
-          return yearPrefix + ctx.getString(R.string.grouping_week) + " " + groupSecond + weekRange;
-        case MONTH:
-          int monthStarts = Integer.parseInt(PrefKey.GROUP_MONTH_STARTS.getString("1"));
-          cal = Calendar.getInstance();
-          if (monthStarts == 1) {
-            cal.set(groupYear, groupSecond - 1, 1);
-            //noinspection SimpleDateFormat
-            return new SimpleDateFormat("MMMM y").format(cal.getTime());
-          } else {
-            dateformat = android.text.format.DateFormat.getLongDateFormat(ctx);
-            int beginYear = groupYear, beginMonth = groupSecond - 1;
-            cal = Calendar.getInstance();
-            cal.set(beginYear, beginMonth, 1);
-            if (cal.getActualMaximum(Calendar.DAY_OF_MONTH) < monthStarts) {
-              cal.set(beginYear, beginMonth + 1, 1);
-            } else {
-              cal.set(Calendar.DATE, monthStarts);
-            }
-            String startDate = dateformat.format(cal.getTime());
-            int endYear = beginYear, endMonth = beginMonth + 1;
-            if (endMonth > Calendar.DECEMBER) {
-              endMonth = Calendar.JANUARY;
-              endYear++;
-            }
-            cal = Calendar.getInstance();
-            cal.set(endYear, endMonth, 1);
-            if (cal.getActualMaximum(Calendar.DAY_OF_MONTH) < monthStarts - 1) {
-              cal.set(Calendar.DATE, cal.getActualMaximum(Calendar.DAY_OF_MONTH));
-            } else {
-              cal.set(Calendar.DATE, monthStarts - 1);
-            }
-            String endDate = dateformat.format(cal.getTime());
-            String monthRange = " (" + startDate + " - " + endDate + " )";
-            return monthRange;
-          }
-        case YEAR:
-          return String.valueOf(groupYear);
-        default:
-          return null;
-      }
-    }
-
-    public static final String JOIN;
-
-    static {
-      JOIN = Utils.joinEnum(Grouping.class);
-    }
-  }
+  public AccountType type;
 
   public Grouping grouping;
-
-  /**
-   * @see <a href="http://www.currency-iso.org/dl_iso_table_a1.xml">http://www.currency-iso.org/dl_iso_table_a1.xml</a>
-   */
-  public enum CurrencyEnum {
-    AFN("Afghani"),
-    ALL("Albania Lek"),
-    DZD("Algerian Dinar"),
-    AOA("Angola Kwanza"),
-    ARS("Argentine Peso"),
-    AMD("Armenian Dram"),
-    AWG("Aruban Florin"),
-    AUD("Australian Dollar"),
-    AZN("Azerbaijanian Manat"),
-    BSD("Bahamian Dollar"),
-    BHD("Bahraini Dinar"),
-    BDT("Bangladesh Taka"),
-    BBD("Barbados Dollar"),
-    BYR("Belarussian Ruble"),
-    BZD("Belize Dollar"),
-    BMD("Bermudian Dollar"),
-    BTN("Bhutan Ngultrum"),
-    BOB("Bolivia Boliviano"),
-    BAM("Bosnia and Herzegovina Convertible Mark"),
-    BWP("Botswana Pula"),
-    BRL("Brazilian Real"),
-    BND("Brunei Dollar"),
-    BGN("Bulgarian Lev"),
-    BIF("Burundi Franc"),
-    KHR("Cambodia Riel"),
-    CAD("Canadian Dollar"),
-    CVE("Cape Verde Escudo"),
-    KYD("Cayman Islands Dollar"),
-    XOF("CFA Franc BCEAO"),
-    XAF("CFA Franc BEAC"),
-    XPF("CFP Franc"),
-    CLP("Chilean Peso"),
-    CNY("China Yuan Renminbi"),
-    COP("Colombian Peso"),
-    KMF("Comoro Franc"),
-    CDF("Congolese Franc"),
-    CRC("Costa Rican Colon"),
-    HRK("Croatian Kuna"),
-    CUP("Cuban Peso"),
-    CUC("Cuba Peso Convertible"),
-    CZK("Czech Koruna"),
-    DKK("Danish Krone"),
-    DJF("Djibouti Franc"),
-    DOP("Dominican Peso"),
-    XCD("East Caribbean Dollar"),
-    EGP("Egyptian Pound"),
-    SVC("El Salvador Colon"),
-    ERN("Eritrea Nakfa"),
-    ETB("Ethiopian Birr"),
-    EUR("Euro"),
-    FKP("Falkland Islands Pound"),
-    FJD("Fiji Dollar"),
-    GMD("Gambia Dalasi"),
-    GEL("Georgia Lari"),
-    GHS("Ghana Cedi"),
-    GIP("Gibraltar Pound"),
-    GTQ("Guatemala Quetzal"),
-    GNF("Guinea Franc"),
-    GYD("Guyana Dollar"),
-    HTG("Haiti Gourde"),
-    HNL("Honduras Lempira"),
-    HKD("Hong Kong Dollar"),
-    HUF("Hungary Forint"),
-    ISK("Iceland Krona"),
-    INR("Indian Rupee"),
-    IDR("Indonesia Rupiah"),
-    IRR("Iranian Rial"),
-    IQD("Iraqi Dinar"),
-    ILS("Israeli Sheqel"),
-    JMD("Jamaican Dollar"),
-    JPY("Japan Yen"),
-    JOD("Jordanian Dinar"),
-    KZT("Kazakhstan Tenge"),
-    KES("Kenyan Shilling"),
-    KRW("Korea Won"),
-    KWD("Kuwaiti Dinar"),
-    KGS("Kyrgyzstan Som"),
-    LAK("Lao Kip"),
-    LVL("Latvian Lats"),
-    LBP("Lebanese Pound"),
-    LSL("Lesotho Loti"),
-    LRD("Liberian Dollar"),
-    LYD("Libyan Dinar"),
-    LTL("Lithuanian Litas"),
-    MOP("Macao Pataca"),
-    MKD("Macedonia Denar"),
-    MGA("Malagasy Ariary"),
-    MWK("Malawi Kwacha"),
-    MYR("Malaysian Ringgit"),
-    MVR("Maldives Rufiyaa"),
-    MRO("Mauritania Ouguiya"),
-    MUR("Mauritius Rupee"),
-    MXN("Mexican Peso"),
-    MDL("Moldovan Leu"),
-    MNT("Mongolia Tugrik"),
-    MAD("Moroccan Dirham"),
-    MZN("Mozambique Metical"),
-    MMK("Myanmar Kyat"),
-    NAD("Namibia Dollar"),
-    NPR("Nepalese Rupee"),
-    ANG("Netherlands Antillean Guilder"),
-    NZD("New Zealand Dollar"),
-    NIO("Nicaragua Cordoba Oro"),
-    NGN("Nigeria Naira"),
-    KPW("North Korean Won"),
-    NOK("Norwegian Krone"),
-    OMR("Omani Rial"),
-    PKR("Pakistan Rupee"),
-    PAB("Panama Balboa"),
-    PGK("Papua New Guinea Kina"),
-    PYG("Paraguay Guarani"),
-    PEN("Peru Nuevo Sol"),
-    PHP("Philippine Peso"),
-    PLN("Poland Zloty"),
-    QAR("Qatari Rial"),
-    RON("Romanian Leu"),
-    RUB("Russian Ruble"),
-    RWF("Rwanda Franc"),
-    SHP("Saint Helena Pound"),
-    WST("Samoa Tala"),
-    STD("Sao Tome and Principe Dobra"),
-    SAR("Saudi Riyal"),
-    RSD("Serbian Dinar"),
-    SCR("Seychelles Rupee"),
-    SLL("Sierra Leone Leone"),
-    SGD("Singapore Dollar"),
-    SBD("Solomon Islands Dollar"),
-    SOS("Somali Shilling"),
-    ZAR("South Africa Rand"),
-    SSP("South Sudanese Pound"),
-    LKR("Sri Lanka Rupee"),
-    SDG("Sudanese Pound"),
-    SRD("Surinam Dollar"),
-    SZL("Swaziland Lilangeni"),
-    SEK("Swedish Krona"),
-    CHF("Swiss Franc"),
-    SYP("Syrian Pound"),
-    TWD("Taiwan Dollar"),
-    TJS("Tajikistan Somoni"),
-    TZS("Tanzanian Shilling"),
-    THB("Thai Baht"),
-    TOP("Tonga Pa’anga"),
-    TTD("Trinidad and Tobago Dollar"),
-    TND("Tunisian Dinar"),
-    TRY("Turkish Lira"),
-    TMT("Turkmenistan New Manat"),
-    AED("UAE Dirham"),
-    UGX("Uganda Shilling"),
-    UAH("Ukraine Hryvnia"),
-    GBP("United Kingdom Pound Sterling"),
-    UYU("Uruguayo Peso"),
-    USD("US Dollar"),
-    UZS("Uzbekistan Sum"),
-    VUV("Vanuatu Vatu"),
-    VEF("Venezuela Bolivar Fuerte"),
-    VND("Vietnam Dong"),
-    YER("Yemeni Rial"),
-    ZMW("Zambian Kwacha"),
-    ZWL("Zimbabwe Dollar"),
-    XXX("No currency "),
-    XAU("Gold"),
-    XPD("Palladium"),
-    XPT("Platinum"),
-    XAG("Silver");
-    private String description;
-
-    CurrencyEnum(String description) {
-      this.description = description;
-    }
-
-    @TargetApi(Build.VERSION_CODES.KITKAT)
-    public String toString() {
-      if (Utils.hasApiLevel(Build.VERSION_CODES.KITKAT)) {
-        try {
-          return Currency.getInstance(name()).getDisplayName();
-        } catch (IllegalArgumentException e) {
-        }
-      }
-      return description;
-    }
-
-    public static CurrencyEnum[] sortedValues() {
-      CurrencyEnum[] result = values();
-      final Collator collator = Collator.getInstance();
-      if (Utils.hasApiLevel(Build.VERSION_CODES.KITKAT)) {
-        Arrays.sort(result, new Comparator<CurrencyEnum>() {
-          @Override
-          public int compare(CurrencyEnum lhs, CurrencyEnum rhs) {
-            int classCompare = Utils.compare(lhs.sortClass(), rhs.sortClass());
-            return classCompare == 0 ?
-                collator.compare(lhs.toString(), rhs.toString()) : classCompare;
-          }
-        });
-      }
-      return result;
-    }
-
-    private int sortClass() {
-      switch (this) {
-        case XXX:
-          return 3;
-        case XAU:
-        case XPD:
-        case XPT:
-        case XAG:
-          return 2;
-        default:
-          return 1;
-      }
-    }
-  }
 
   public static final int DEFAULT_COLOR = 0xff009688;
 
@@ -678,34 +240,6 @@ public class Account extends Model {
     this("", (long) 0, "");
   }
 
-  public static Currency getLocalCurrency() {
-    Currency result = null;
-    TelephonyManager telephonyManager = (TelephonyManager) MyApplication.getInstance()
-        .getSystemService(Context.TELEPHONY_SERVICE);
-    if (telephonyManager != null) {
-      try {
-        String userCountry = telephonyManager.getNetworkCountryIso();
-        if (TextUtils.isEmpty(userCountry)) {
-          userCountry = telephonyManager.getSimCountryIso();
-        }
-        if (!TextUtils.isEmpty(userCountry)) {
-          result = Utils.getSaveInstance(Currency.getInstance(new Locale("", userCountry)));
-        }
-      } catch (Exception e) {
-        //fall back to currency from locale
-      }
-    }
-    if (result == null) {
-      try {
-        //makeSure we know about the currency
-        result = Utils.getSaveInstance(Currency.getInstance(Locale.getDefault()));
-      } catch (IllegalArgumentException e) {
-        result = Currency.getInstance("EUR");
-      }
-    }
-    return result;
-  }
-
   /**
    * Account with currency from locale, of type CASH and with DEFAULT_COLOR
    *
@@ -714,11 +248,11 @@ public class Account extends Model {
    * @param description    the description
    */
   public Account(String label, long openingBalance, String description) {
-    this(label, getLocalCurrency(), openingBalance, description, Type.CASH, DEFAULT_COLOR);
+    this(label, Utils.getLocalCurrency(), openingBalance, description, AccountType.CASH, DEFAULT_COLOR);
   }
 
   public Account(String label, Currency currency, long openingBalance, String description,
-                 Type type, int color) {
+                 AccountType type, int color) {
     this.label = label;
     this.currency = currency;
     this.openingBalance = new Money(currency, openingBalance);
@@ -751,9 +285,9 @@ public class Account extends Model {
     this.openingBalance = new Money(this.currency,
         c.getLong(c.getColumnIndexOrThrow(KEY_OPENING_BALANCE)));
     try {
-      this.type = Type.valueOf(c.getString(c.getColumnIndexOrThrow(KEY_TYPE)));
+      this.type = AccountType.valueOf(c.getString(c.getColumnIndexOrThrow(KEY_TYPE)));
     } catch (IllegalArgumentException ex) {
-      this.type = Type.CASH;
+      this.type = AccountType.CASH;
     }
     try {
       this.grouping = Grouping.valueOf(c.getString(c.getColumnIndexOrThrow(KEY_GROUPING)));
