+++ myExpenses/src/main/java/org/totschnig/myexpenses/model/AccountGrouping.java
@@ -1,0 +1,8 @@
+package org.totschnig.myexpenses.model;
+
+/**
+ * grouping of accounts in account list
+ */
+public enum AccountGrouping {
+  NONE, TYPE, CURRENCY
+}
